refactor: action 상수화 및 파일명 변경(reduce->reducer)
