--- src/context/constant.jsx
@@ -1,10 +1,0 @@
-const initialCardInfo = {
-  company: '',
-  cardNumber: { first: '', second: '', third: '', fourth: '' },
-  dueDate: { month: '', year: '' },
-  owner: '',
-  cvc: '',
-  password: { first: '', second: '' },
-}
-
-export { initialCardInfo }
