+++ src/context/reducer.jsx
@@ -1,15 +1,20 @@
-import { initialCardInfo } from 'context/constant'
+import {
+  initialCardInfo,
+  CHANGE_CARD_COMPANY,
+  CHANGE_CARD_INFO,
+  CLEAR_CARD_INFO,
+} from 'context/cardInfo-context'
 
 const reducer = (state, action) => {
   const { type, name, key, value } = action
 
   switch (type) {
-    case 'CHANGE_CARD_COMPANY':
+    case CHANGE_CARD_COMPANY:
       return {
         ...state,
         [name]: value,
       }
-    case 'CHANGE_CARD_INFO':
+    case CHANGE_CARD_INFO:
       if (key) {
         return {
           ...state,
@@ -20,7 +25,7 @@ const reducer = (state, action) => {
         }
       }
       return { ...state, [name]: value }
-    case 'CLEAR_CARD_INFO':
+    case CLEAR_CARD_INFO:
       return initialCardInfo
     default:
       return state
